move shareIssue
